Reorganised the code for our main window and splash screen window.
